fixed dp bug on demand range

+++ src/idinn/dual_controller/dynamic_programming.py
@@ -5,6 +5,8 @@ import torch
 from numba import njit, types
 from numba.typed import Dict, List
 
+from .. import demand
+
 
 class DynamicProgrammingController(BaseDualController):
     def __init__(self) -> None:
@@ -92,7 +94,8 @@ class DynamicProgrammingController(BaseDualController):
         max_order = max_ip + min_ip
         dim_pipeline = lr - le - 1
 
-        demand_prob = Dict.empty(key_type=types.float64, value_type=types.float64)
+        demand_prob = Dict.empty(key_type=types.int64,
+                                 value_type=types.float64)
         demand_prob_ = sourcing_model.demand_generator.enumerate_support()
         for k, v in demand_prob_.items():
             demand_prob[k] = v
